`generate` does not create token on server
- move create utils into `create` package
