+++ pkg/gardenadm/cmd/token/generate/options.go
@@ -8,31 +8,26 @@ import (
 	"github.com/spf13/pflag"
 
 	"github.com/gardener/gardener/pkg/gardenadm/cmd"
-	tokenutils "github.com/gardener/gardener/pkg/gardenadm/cmd/token/utils"
 )
 
 // Options contains options for this command.
 type Options struct {
 	*cmd.Options
-	// CreateOptions are the options for creating a bootstrap token.
-	CreateOptions *tokenutils.Options
 }
 
 // ParseArgs parses the arguments to the options.
-func (o *Options) ParseArgs(args []string) error {
-	return o.CreateOptions.ParseArgs(args)
+func (o *Options) ParseArgs(_ []string) error {
+	return nil
 }
 
 // Validate validates the options.
 func (o *Options) Validate() error {
-	return o.CreateOptions.Validate()
+	return nil
 }
 
 // Complete completes the options.
 func (o *Options) Complete() error {
-	return o.CreateOptions.Complete()
+	return nil
 }
 
-func (o *Options) addFlags(fs *pflag.FlagSet) {
-	o.CreateOptions.AddFlags(fs)
-}
+func (o *Options) addFlags(_ *pflag.FlagSet) {}
